Move local queries commands to separate file
